requirements.txt
Django==1.9
simplejson
django-uuidfield
